// Libreria-Online/back/src/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EstadoOferta = {
  ESPERANDO_ADMIN: 'ESPERANDO_ADMIN',
  ESPERANDO_PROVEEDOR: 'ESPERANDO_PROVEEDOR',
  ACEPTADA: 'ACEPTADA',
  RECHAZADA: 'RECHAZADA'
} as const

export type EstadoOferta = (typeof EstadoOferta)[keyof typeof EstadoOferta]


export const CategoriaLibro = {
  FICCION: 'FICCION',
  FANTASIA: 'FANTASIA',
  TERROR: 'TERROR',
  HISTORIA: 'HISTORIA',
  INFANTIL: 'INFANTIL',
  GENERAL: 'GENERAL'
} as const

export type CategoriaLibro = (typeof CategoriaLibro)[keyof typeof CategoriaLibro]